+++ IO/ANALOG_SENSORS/PRESSURE_SENSORS/FLEXIFORCE_25LB/FLEXIFORCE_25LB.py
@@ -1,0 +1,7 @@
+from flojoy import flojoy, OrderedPair, DataContainer
+from typing import Optional
+
+
+@flojoy(deps={"pyserial": "3.5"})
+def FLEXIFORCE_25LB(default: Optional[DataContainer] = None) -> OrderedPair:
+    return OrderedPair()
