+++ IO/ANALOG_SENSORS/Thermocouples/LM34/LM34.py
@@ -1,0 +1,7 @@
+from flojoy import flojoy, OrderedPair, DataContainer
+from typing import Optional
+
+
+@flojoy()
+def LM34(default: Optional[DataContainer] = None) -> OrderedPair:
+    return OrderedPair()
