--- IO/Analog_Sensors/Pressure_Sensors/FLEXIFORCE_25LB/FLEXIFORCE_25LB.py
@@ -1,6 +1,0 @@
-from flojoy import flojoy, OrderedPair, DataContainer
-from typing import Optional
-
-
-def FLEXIFORCE_25LB(default: Optional[DataContainer] = None) -> OrderedPair:
-    return OrderedPair()
