--- IO/Analog_Sensors/Thermocouples/LM34/LM34.py
@@ -1,6 +1,0 @@
-from flojoy import flojoy, OrderedPair, DataContainer
-from typing import Optional
-
-
-def LM34(default: Optional[DataContainer] = None) -> OrderedPair:
-    return OrderedPair()
