+++ IO/INSTRUMENTS/DAQ_BOARDS/NI/COMING_FALL_2023/COMING_FALL_2023.py
@@ -1,0 +1,7 @@
+from flojoy import flojoy, OrderedPair, DataContainer
+from typing import Optional
+
+
+@flojoy()
+def COMING_FALL_2023(default: Optional[DataContainer] = None) -> OrderedPair:
+    return OrderedPair()
