+++ IO/INSTRUMENTS/MOCK/WEINSCHEL8320/Weinschel_8320.yaml
