+++ IO/INSTRUMENTS/SOURCEMETERS/KEITHLEY/24XX/BASIC/IV_SWEEP/IV_SWEEP.py
@@ -1,0 +1,79 @@
+from flojoy import flojoy, OrderedPair, Vector, node_initialization, NodeInitContainer
+import serial
+
+
+@flojoy(deps={"pyserial": "3.5"})
+def IV_SWEEP(
+    init_container: NodeInitContainer, default: OrderedPair | Vector
+) -> OrderedPair:
+    """The KEITHLEY2400 node takes a IV curve measurement with a Keithley 2400 source meter, send voltages, and measures currents.
+
+    Parameters
+    -----------
+    comport : string
+         defines the serial communication port for the Keithley2400 source meter.
+    baudrate : float
+         specifies the baud rate for the serial communication between the Keithley2400 and the computer
+    """
+
+    # Start serial communication with the instrument
+    # ser: serial = serial.Serial()
+
+    ser = init_container.get()
+    if ser is None:
+        raise ValueError("Serial communication is not open")
+
+    # Keithley 2400 Configuration
+    ser.write(b"*RST\n")  # reinitialisation of the instrument
+    ser.write(b":SOUR:FUNC:MODE VOLT\n")  # Sourcing tension
+    ser.write(b':SENS:FUNC "CURR"\n')  # Measuring current
+    ser.write(
+        b":SENS:CURR:PROT 1.05\n"
+    )  # Current protection set at 1.05A (Keithley 2400)
+
+    match default:
+        case OrderedPair():
+            voltages = default.y
+        case Vector():
+            voltages = default.v
+
+    currents_neg: list[float] = []  # measured currents
+
+    for voltage in voltages:
+        ser.write(b":SOUR:VOLT %f\n" % voltage)  # Source Tension (V)
+        ser.write(b":OUTP ON\n")  # Instrument output open
+        ser.write(b":INIT\n")  # Start measuring
+        ser.write(b":FETC?\n")  # Retrieve the measured values
+
+        current_str: str = (
+            ser.readline().decode("ascii").strip()
+        )  # Save answers in a string
+        voltage_current_values: str = current_str.split(
+            ",")  # Split the string
+        currents_neg.append(-float(voltage_current_values[1]))
+
+        ser.write(b":OUTP OFF\n")  # Close output from Instrument
+
+    # Close Serial Communication
+    ser.close()
+
+    return OrderedPair(x=voltages, y=currents_neg)
+
+
+@node_initialization(for_node=IV_SWEEP)
+def init(comport: str = "/dev/ttyUSB0", baudrate: float = 9600):
+    ser: serial = serial.Serial()
+
+    # Specific parameters
+    ser.port = comport  # Specify serial port for com
+    ser.baudrate = baudrate  # Specify Baudrate
+
+    # General parameters
+    ser.bytesize = serial.EIGHTBITS  # Specify Bites number
+    ser.parity = serial.PARITY_NONE  # Specify Parity
+    ser.stopbits = serial.STOPBITS_ONE  # Specify Stop bites
+    ser.timeout = 1
+    # Open Serial Com
+    ser.open()
+
+    return ser
