+++ IO/MICROCONTROLLERS/ESP32/ESP32.py
@@ -1,0 +1,7 @@
+from flojoy import flojoy, OrderedPair, DataContainer
+from typing import Optional
+
+
+@flojoy()
+def ESP32(default: Optional[DataContainer] = None) -> OrderedPair:
+    return OrderedPair()
