+++ IO/MICROCONTROLLERS/RPI2040/RPI2040.py
@@ -1,0 +1,7 @@
+from flojoy import flojoy, OrderedPair, DataContainer
+from typing import Optional
+
+
+@flojoy()
+def RPI2040(default: Optional[DataContainer] = None) -> OrderedPair:
+    return OrderedPair()
