+++ IO/MICROCONTROLLERS/STM32/STM32.py
@@ -1,0 +1,7 @@
+from flojoy import flojoy, OrderedPair, DataContainer
+from typing import Optional
+
+
+@flojoy()
+def STM32(default: Optional[DataContainer] = None) -> OrderedPair:
+    return OrderedPair()
