+++ IO/MOTION/MOTOR_DRIVER/BLDC/LINENGINEERING/LINENGINEERING.py
@@ -1,0 +1,7 @@
+from flojoy import flojoy, OrderedPair, DataContainer
+from typing import Optional
+
+
+@flojoy()
+def LINENGINEERING(default: Optional[DataContainer] = None) -> OrderedPair:
+    return OrderedPair()
