+++ IO/MOTION/MOTOR_DRIVER/BLDC/TINYMOVR/TINYMOVR.py
@@ -1,0 +1,7 @@
+from flojoy import flojoy, OrderedPair, DataContainer
+from typing import Optional
+
+
+@flojoy()
+def TINYMOVR(default: Optional[DataContainer] = None) -> OrderedPair:
+    return OrderedPair()
