+++ IO/MOTION/MOTOR_DRIVER/SERVO/TINYMOVR2/TINYMOVR2.py
@@ -1,0 +1,7 @@
+from flojoy import flojoy, OrderedPair, DataContainer
+from typing import Optional
+
+
+@flojoy()
+def TINYMOVR2(default: Optional[DataContainer] = None) -> OrderedPair:
+    return OrderedPair()
