+++ IO/MOTION/MOTOR_DRIVER/STEPPER/ADAFRUIT/ADAFRUIT.py
@@ -1,0 +1,7 @@
+from flojoy import flojoy, OrderedPair, DataContainer
+from typing import Optional
+
+
+@flojoy()
+def ADAFRUIT(default: Optional[DataContainer] = None) -> OrderedPair:
+    return OrderedPair()
