+++ IO/MOTION/MOTOR_DRIVER/STEPPER/ZABER/ZABER.py
@@ -1,0 +1,7 @@
+from flojoy import flojoy, OrderedPair, DataContainer
+from typing import Optional
+
+
+@flojoy()
+def ZABER(default: Optional[DataContainer] = None) -> OrderedPair:
+    return OrderedPair()
