+++ IO/PROTOCOLS/SCPI/SCPI.py
@@ -1,0 +1,7 @@
+from flojoy import flojoy, OrderedPair, DataContainer
+from typing import Optional
+
+
+@flojoy()
+def SCPI(default: Optional[DataContainer] = None) -> OrderedPair:
+    return OrderedPair()
