+++ IO/PROTOCOLS/SERIAL/BASIC/TIMESERIES1/TIMESERIES1.py
